requirements.txt
streamlit==1.28.1
PyMuPDF==1.23.5
openai==0.28.1
faiss-cpu==1.7.4
Pillow==10.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3